requirements.txt
discord.py>=2.0.0
psutil
pyyaml
requests
pywin32